### 2022-07-18 #4578 @editorialbot
[PRE REVIEW]: geostan: An R package for Bayesian spatial analysis
**Submitting author:** @ConnorDonegan (<a href="http://orcid.org/0000-0002-9698-5443">Connor Donegan</a>)
**Repository:** https://github.com/ConnorDonegan/geostan
**Branch with paper.md** (empty if default branch): JOSS
**Version:** V0.3.0
