### 2022-07-27 #4578 @Kevin-Mattheus-Moerman
@ConnorDonegan I am an AEiC with JOSS and here to help with initial steps before we proceed. Can you check the following:

- [ ] JOSS requires an open source license. Please add an [OSI approved open source license](https://opensource.org/licenses) to your software. It is recommended to do so in the form of a plain text `LICENSE` file placed in the main folder of your repository.
- [ ] Please check those potentially missing DOIs :point_up:, you can call `@editorialbot check references` to check the references again.
