### 2022-07-28 188f5cb
added missing DOIs
