+++ paper.bib
@@ -1,11 +1,3 @@
-@article{student_1914,
-  title={The elimination of spurious correlation due to position in time and space},
-  author={{``Student'' [W.S. Gausset]}},
-  journal={Biometrika},
-  volume={10},
-  year={1914},
-  pages={179-180}
-}
 
 @article{donegan_2022,
   title={Building spatial conditional autoregressive (CAR) models in the {S}tan programming language},
@@ -22,7 +14,8 @@
   pages={100450},
   year={2020},
   volume={38},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.spasta.2020.100450}
 }
 
 @article{donegan_2021,
@@ -85,7 +78,8 @@
   volume={95},
   number={4},
   pages={740--760},
-  year={2005}
+  year={2005},
+  doi={10.1111/j.1467-8306.2005.00484.x}
 }
 
 @book{griffith_2019,
@@ -93,7 +87,8 @@
   author={Griffith, Daniel and Chun, Yongwan and Li, Bin},
   year={2019},
   publisher={Academic Press},
-  address={London}
+  address={London},
+  doi={10.1016/C2017-0-01015-7}
 }
 
 @book{haining_2020,
@@ -111,7 +106,8 @@
   journal={Geographical Analysis},
   volume={39},
   number={4},
-  pages={357--375}
+  pages={357--375},
+  doi={10.1111/j.1538-4632.2007.00708.x}
 }
 
 @article{bazuin_2013,
@@ -151,7 +147,8 @@
   volume={53},
   issue={8},
   year={2009},
-  pages={3016-3032}
+  pages={3016-3032},
+  doi={10.1016/j.csda.2008.07.033}
 }
 
 @article{bernardinelli_1997,
@@ -161,7 +158,8 @@
   journal={Statistics in Medicine},
   volume={16},
   number={7},
-  pages={741--752}
+  pages={741--752},
+  doi={10.1002/(sici)1097-0258(19970415)16:7<741::aid-sim501>3.0.co;2-1}
 }
 
 @article{xia_1998,
@@ -172,7 +170,8 @@
   number={18},
   pages={2025--2043},
   year={1998},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/(sici)1097-0258(19980930)17:18<2025::aid-sim865>3.0.co;2-m}
 }
 
 @article{besag_1974,
@@ -182,7 +181,8 @@
   volume={36},
   number={2},
   pages={192-225},
-  year={1974}
+  year={1974},
+  doi={10.1111/j.2517-6161.1974.tb00999.x}
 }
 @article{besag_1991,
   title={Bayesian image restoration, with two applications in spatial statistics},
@@ -191,7 +191,8 @@
   volume={43},
   issue={1},
   pages={1-20},
-  year={1991}
+  year={1991},
+  doi={10.1007/BF00116466}
 }
 
 @article{riebler_2016,
@@ -202,7 +203,8 @@
   number={4},
   pages={1145--1165},
   year={2016},
-  publisher={Sage Publications Sage UK: London, England}
+  publisher={Sage Publications Sage UK: London, England},
+  doi={10.1177/0962280216660421}
 }
 
 @article{morris_2019,
@@ -212,7 +214,8 @@
   volume={31},
   pages={100301},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.sste.2019.100301}
 }
 
 @article{gabry_2019,
@@ -231,7 +234,7 @@
     title = {Applied spatial data analysis with {R}, Second edition},
     year = {2013},
     publisher = {Springer, NY},
[…]
