+++ paper.bib
@@ -332,7 +332,8 @@
   number={2},
   pages={93-115},
   year={1995},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1538-4632.1995.tb00338.x}
 }
 
 
