### 2022-08-13 095a649
aple function moved to spatialreg, external
