### 2022-08-21 #4578 @jbytecode
👋👋👋 Dear @Volkerschmid,  @becarioprecario,  @wcjochem  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**geostan: An R package for Bayesian spatial analysis**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4578).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a seperate thread. In that thread, there will be 20 check items for each single reviewer.

Thank you in advance!
