### 2022-08-21 #4578 @jbytecode
@ConnorDonegan - Thank you for submitting your paper for JOSS. I am the handling editor.

The reviewers you suggested seem to be inactive for a while. I will wait for a couple of days to get a response (Because it is the vacation time for much of us).

Thank you in advance.
