### 2022-08-22 #4578 @wcjochem
Dear @jbytecode this sounds like a very interesting submission. Unfortunately, I'm working on another JOSS review currently and with other projects deadlines and upcoming leave I wouldn't be available to help to start with this until mid-September.
