### 2022-08-28 #4578 @cole-brokamp
Hi @jbytecode —- I don’t feel like I would be able to give a quality review here.  Although I work with place-based data a lot, I don’t do any spatial analyses and I’ve never done Bayesian modeling nor used Stan.
