### 2022-08-28 #4578 @becarioprecario
Hi,

I can review this paper.

Best,

Virgilio
________________________________
De: Mehmet Hakan Satman ***@***.***>
Enviado: domingo, 28 de agosto de 2022 15:39
Para: openjournals/joss-reviews ***@***.***>
Cc: Virgilio Gómez Rubio ***@***.***>; Mention ***@***.***>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: geostan: An R package for Bayesian spatial analysis (Issue #4578)

@cole-brokamp<https://github.com/cole-brokamp> - thank you for the response. Hope we can work on future projects.
