### 2022-09-12 d64ac2e
latex math in docs
