+++ R/stan_glm.R
@@ -60,78 +60,93 @@
 #'
 #' ### Poisson models and disease mapping
 #' 
-#' In spatial statistics, Poisson models are often used to calculate incidence rates (mortality rates, or disease incidence rates) for administrative areas like counties or census tracts. If `Y` are counts of cases, and `P` are populations at risk, then the crude rates are `Y/P`. The purpose is to model risk, `eta`, for which crude rates are a (noisy) indicator. Our analysis should also respect the fact that the amount of information contained in the observations, `Y/P`, increases with `P`. Hierarchical Poisson models are often the best way to incorporate all of this information.
+#' In spatial statistics, Poisson models are often used to calculate incidence rates (mortality rates, or disease incidence rates) for administrative areas like counties or census tracts. If \eqn{y} are counts of cases, and \eqn{P} are populations at risk, then the crude rates are \eqn{y/P}. The purpose is to model risk \eqn{\eta} for which crude rates are a (noisy) indicator. Our analysis should also respect the fact that the amount of information contained in the observations \eqn{y/P} increases with \eqn{P}. Hierarchical Poisson models are often used to incorporate all of this information.
 #'
-#' For the Poisson model, `Y` is specified as the outcome and the log of the population at risk, `log(P)`, needs to be provided as an offset term. For such a case, disease incidence across the collection of areas could be modeled as:
-#' ```
-#' Y ~ Poisson(exp(log(P) + eta))
-#' eta = alpha + A
-#' A ~ Guass(0, tau)
-#' tau ~ student(20, 0, 2),
-#' ```
-#' where `alpha` is the mean log-risk (incidence rate) and `A` is a vector of (so-called) random effects, which enable partial pooling of information across observations. Covariates can be added to the model for the log-rates, such that `eta = alpha + X * beta + A`. See the example section of this document for a demonstration (where the denominator of the outcome is the expected count, rather than population at risk).
+#' For the Poisson model, \eqn{y} is specified as the outcome and the log of the population at risk `log(P)` needs to be provided as an offset term. For such a case, disease incidence across the collection of areas could be modeled as:
+#' \deqn{
+#' y \sim Poisson(e^{log(P) + \eta}) \\
+#' \eta = \alpha + A \\
+#' A \sim Guass(0, \tau) \\
+#' \tau \sim student(20, 0, 2),
+#' }
+#' where \eqn{\alpha} is the mean log-risk (incidence rate) and \eqn{A} is a vector of (so-called) random effects, which enable partial pooling of information across observations. Covariates can be added to the model for the log-rates, such that \eqn{\eta = \alpha + X * \beta + A}. See the example section of this document for a demonstration (where the denominator of the outcome is the expected count, rather than population at risk).
 #'
 #' Note that the denominator for the rates is specified as a log-offset to provide a consistent, formula-line interface to the model. An equivalent, and perhaps more intuitive, specification is the following:
-#' ```
-#' Y ~ Poisson(P * exp(eta))
-#' ```
-#' where `P` is still the population at risk and `exp(eta)` is the incidence rate (risk). 
+#' \deqn{
+#' y \sim Poisson(P * e^{\eta})
+#' }
[…]
