+++ R/stan_icar.R
@@ -55,49 +55,49 @@
 #' @param ... Other arguments passed to \link[rstan]{sampling}. For multi-core processing, you can use \code{cores = parallel::detectCores()}, or run \code{options(mc.cores = parallel::detectCores())} first.
 #' @details
 #' 
-#'  The Stan code for the ICAR component of the model and the BYM2 option is from Morris et al. (2019) with adjustments to enable non-binary weights and disconnected graph structures (see Freni-Sterrantino (2018) and Donegan (2021)).
+#'  The intrinsic conditional autoregressive (ICAR) model for spatial data was introduced by Besag et al. (1991). The Stan code for the ICAR component of the model and the BYM2 option is from Morris et al. (2019) with adjustments to enable non-binary weights and disconnected graph structures (see Freni-Sterrantino (2018) and Donegan (2021)).
 #'
 #' The exact specification depends on the `type` argument. 
 #'
 #' ### 'icar'
 #'
 #' For Poisson models for count data, y, the basic model specification (`type = "icar"`) is:
-#' ```
-#' y ~ Poisson(exp(offset + mu + phi))
-#' phi ~ ICAR(spatial_scale)
-#' spatial_scale ~ Gaussian(0, 1)
-#' ```
-#'  where `mu` contains an intercept and potentially covariates. The spatial trend, `phi`, has a mean of zero and a single scale parameter, `spatial_scale`.
+#' \deqn{
+#' y ~ Poisson(e^{O + \mu + \phi}) \\
+#' \phi \sim ICAR(\tau_s) \\
+#' \tau_s \sim Gauss(0, 1)
+#' }
+#'  where \eqn{\mu} contains an intercept and potentially covariates. The spatial trend \eqn{phi} has a mean of zero and a single scale parameter \eqn{\tau_s} (which user's will see printed as the parameter named `spatial_scale`).
 #' 
-#' The ICAR prior model is a CAR model that has a spatial autocorrelation parameter \code{car_alpha} equal to 1 (see \link[geostan]{stan_car}). Thus the ICAR prior places high probability on a smooth spatially (or temporally) varying mean. This is rarely sufficient to model the amount of variation present in social and health data.
+#' The ICAR prior model is a CAR model that has a spatial autocorrelation parameter \eqn{\rho} equal to 1 (see \link[geostan]{stan_car}). Thus the ICAR prior places high probability on a very smooth spatially (or temporally) varying mean. This is rarely sufficient to model the amount of variation present in social and health data.
 #'
 #' ### 'bym'
 #'
-#' Often, an observational-level random effect term, `theta`, is added to capture (heterogeneous or unstructured) deviations from `mu + phi`. The combined term is referred to as a convolution term:
-#' ```
-#'  convolution = phi + theta.
-#' ```
+#' Often, an observational-level random effect term, `theta`, is added to capture (heterogeneous or unstructured) deviations from \eqn{\mu + \phi}. The combined term is referred to as a convolution term:
+#' \eqn{
+#'  \text{convolution} = \phi + \theta.
+#' }
 #' This is known as the BYM model (Besag et al. 1991), and can be specified using `type = "bym"`:
-#' ```
-#' y ~ Poisson(exp(offset + mu + phi + theta))
-#' phi ~ ICAR(spatial_scale)
-#' theta ~ Gaussian(0, theta_scale)
-#' spatial_scale ~ Gaussian(0, 1)
-#' theta_scale ~ Gaussian(0, 1)
-#' ```
+#' \eqn{
+#' y \sim Poisson(e^{O + \mu + \phi + \theta}) \\
+#' \phi \sim ICAR(\tau_s) \\
+#' \theta \sim Gaussian(0, \tau_{ns})
+#' \tau_s \sim Gaussian(0, 1)
[…]
