+++ R/stan_sar.R
@@ -56,22 +56,30 @@
 #' 
 #' @details
 #'
-#' Discussions of SAR models may be found in Cliff and Ord (1981), Cressie (2015, Ch. 6), LeSage and Pace (2009), and LeSage (2014). The spatial statistics literature conceptualizes the SAR model in relation to the (alternative) conditionally-specified spatial autoregressive model (see \code{\link[geostan]{stan_car}}). The spatial econometrics literature refers to the simultaneously-specified spatial autoregressive (SAR) model as the spatial error model (SEM), contrasting SEM with the spatial lag model (which contains a spatially-lagged dependent variable on the right-hand-side of the regression equation). A valuable property of the SAR model is its highly interpretable spatial autocorrelation parameter, which usually ranges from -1 to 1 (similar to a correlation coefficient). Cressie (2015, 408; also Cressie and Wikle 2011, 200) argues that a drawback of the SAR model is that the errors \eqn{\epsilon} are correlated with the signal \eqn{y}. 
+#' Discussions of SAR models may be found in Cliff and Ord (1981), Cressie (2015, Ch. 6), LeSage and Pace (2009), and LeSage (2014). 
 #'
 #' The general scheme of the SAR model for numeric vector \eqn{y} is
 #' \deqn{
-#' \boldsymbol y = \boldsymbol \mu + (\boldsymbol I - \rho \boldsymbol W)^{-1}\boldsymbol \epsilon \\
-#' \boldsymbol \epsilon \sim Gauss(\boldsymbol 0, \sigma^2 \boldsymbol I)
+#'  y = \mu + ( I - \rho W)^{-1} \epsilon \\
+#' \epsilon \sim Gauss(0, \sigma^2 I)
 #' }
-#' where \eqn{W} is the spatial weights matrix, \eqn{I} is the n-by-n identity matrix, and \eqn{\rho} is a spatial autocorrelation parameter. In words, the errors of the regression equation are spatially autocorrelated. A common interpretation is that there are causes of \eqn{y} that are not included in the model and which combine in such a way as to form spatial patterns. This also describes a situation in which neighboring observations are similar to one another, i.e., they share many attributes. 
+#' where \eqn{W} is the spatial weights matrix, \eqn{I} is the n-by-n identity matrix, and \eqn{\rho} is a spatial autocorrelation parameter. In words, the errors of the regression equation are spatially autocorrelated. 
+#'
+#' Re-arranging terms, the model can also be written as follows:
+#' \deqn{
+#'  y = \mu + \rho W (y - \mu)  + \epsilon
+#' }
+#' which perhaps shows more intuitively the implicit spatial trend component, \eqn{\rho W (y - \mu)}.
 #' 
-#' Most often, this model is applied directly to observations (referred to below as the auto-Gaussian model). The SAR model can also be applied to a vector of parameters inside a hierarchical model. The latter enables spatial autocorrelation to be modeled when the observations are discrete counts (e.g., disease incidence data). 
+#' Most often, this model is applied directly to observations (referred to below as the auto-Gaussian model). The SAR model can also be applied to a vector of parameters inside a hierarchical model. The latter enables spatial autocorrelation to be modeled when the observations are discrete counts (e.g., disease incidence data).
+#'
[…]

[…]
