### 2022-09-13 8df7d0e
latex \text{} fails
