+++ R/stan_icar.R
@@ -75,7 +75,7 @@
 #'
 #' Often, an observational-level random effect term, `theta`, is added to capture (heterogeneous or unstructured) deviations from \eqn{\mu + \phi}. The combined term is referred to as a convolution term:
 #' \eqn{
-#'  \text{convolution} = \phi + \theta.
+#'  convolution = \phi + \theta.
 #' }
 #' This is known as the BYM model (Besag et al. 1991), and can be specified using `type = "bym"`:
 #' \eqn{
@@ -90,7 +90,7 @@
 #' 
 #' Riebler et al. (2016) introduce a variation on the BYM model (`type = "bym2"`). This specification combines \eqn{\phi} and \eqn{\theta} using a mixing parameter \eqn{\rho} that controls the proportion of the variation that is attributable to the spatially autocorrelated term \eqn{\phi} rather than the spatially unstructured term \eqn{\theta}. The terms share a single scale parameter:
 #' \deqn{
-#' \text{convolution} = [sqrt(\rho/\text{scale_factor}) * \tilde{\phi} + sqrt(1 - \rho)  \tilde{\theta}] * \tau_s \\
+#' convolution = [sqrt(\rho * scale_factor) * \tilde{\phi} + sqrt(1 - \rho)  \tilde{\theta}] * \tau_s \\
 #' \tilde{\phi} \sim Gaussian(0, 1) \\
 #' \tilde{\theta} \sim Gaussian(0, 1) \\
 #' \tau_s \sim Gaussian(0, 1)
@@ -205,13 +205,13 @@
 #'
 #' Internally, `geostan` will keep the index values of each censored observation, and the index value of each of the fully observed outcome values. For all observed counts, the likelihood statement will be:
 #' \deqn{
-#' p(y_i | \text{data}, \text{model}) = poisson(y_i | \mu_i), 
+#' p(y_i | data, model) = poisson(y_i | \mu_i), 
 #' }
 #' as usual, where \eqn{\mu_i} may include whatever spatial terms are present in the model.
 #'
 #' For each censored count, the likelihood statement will equal the cumulative Poisson distribution function for values zero through the censor point:
 #' \deqn{
-#' p(y_i | \text{data}, \text{model}) = \sum_{m=0}^{M} Poisson( m | \mu_i),
+#' p(y_i | data, model) = \sum_{m=0}^{M} Poisson( m | \mu_i),
 #' }
 #' where \eqn{M} is the censor point and \eqn{\mu_i} again is the fitted value for the \eqn{i^{th}} observation.
 #' 
