+++ R/stan_car.R
@@ -200,13 +200,13 @@
 #'
 #' Internally, `geostan` will keep the index values of each censored observation, and the index value of each of the fully observed outcome values. For all observed counts, the likelihood statement will be:
 #' \deqn{
-#' p(y_i | \text{data}, \text{model}) = poisson(y_i | \mu_i), 
+#' p(y_i | data, model) = poisson(y_i | \mu_i), 
 #' }
 #' as usual, where \eqn{\mu_i} may include whatever spatial terms are present in the model.
 #'
 #' For each censored count, the likelihood statement will equal the cumulative Poisson distribution function for values zero through the censor point:
 #' \deqn{
-#' p(y_i | \text{data}, \text{model}) = \sum_{m=0}^{M} Poisson( m | \mu_i),
+#' p(y_i | data, model) = \sum_{m=0}^{M} Poisson( m | \mu_i),
 #' }
 #' where \eqn{M} is the censor point and \eqn{\mu_i} again is the fitted value for the \eqn{i^{th}} observation.
 #' 
@@ -261,23 +261,14 @@
 #'                 car_parts = cp,
 #'                 data = georgia,
 #'                 family = poisson(),
-#'                 iter = 800, chains = 2 # for example speed only
+#'                 iter = 800, chains = 1 # for example speed only
 #'                  )
 #' rstan::stan_rhat(fit$stanfit)
 #' rstan::stan_mcse(fit$stanfit)
 #' print(fit)
 #' sp_diag(fit, georgia)
 #'
-#' # censored count outcomes
-#' sum(is.na(georgia$deaths.female))
-#' fit <- stan_car(deaths.female ~ offset(log(pop.at.risk.female)),
-#'                 car_parts = cp,
-#'                 data = georgia,
-#'                 family = poisson(),
-#'                 censor_point = 9,
-#'                 iter = 800, chains = 2 # for example speed only
-#'                 )
-#' 
+#' \donttest{
 #' ## DCAR specification (inverse-distance based)
 #' library(sf)
 #' A <- shape2mat(georgia, "B")
@@ -292,6 +283,7 @@
 #'                iter = 800, chains = 1 # for example speed only 
 #' )
 #' print(fit)
+#' }
 #' 
 #' @export
 #' @md
