### 2022-09-13 679c41a
plot rho for SAR, give news
