+++ R/stan_car.R
@@ -65,9 +65,9 @@
 #' }
 #' where \eqn{I} is the identity matrix, \eqn{\rho} is a spatial dependence parameter, \eqn{C} is a spatial connectivity matrix, and \eqn{M} is a diagonal matrix of variance terms. The diagonal of \eqn{M} contains a scale parameter \eqn{\tau} multiplied by a vector of weights (often set to be proportional to the inverse of the number of neighbors assigned to each site). The CAR model owes its name to the fact that this joint distribution corresponds to a set of conditional distributions that relate the expected value of each observation to a function of neighboring values, i.e., the Markov condition holds:
 #' \deqn{
-#' E(y_i | y_1, y_2, \dots y_{i-1}, y_{i+1}, y_n) = \mu_i + \rho \sum_{j=1}^n c_{i,j} (y_j - \mu_j),
+#' E(y_i | y_1, y_2, \dots, y_{i-1}, y_{i+1}, \dots, y_n) = \mu_i + \rho \sum_{j=1}^n c_{i,j} (y_j - \mu_j),
 #' }
-#' where entries of \eqn{c_{i,j}} are non-zero only if \eqn{j \in N(i)}, and \eqn{N(i)} indexes the sites that are neighbors of the \eqn{i^{th}} site.
+#' where entries of \eqn{c_{i,j}} are non-zero only if \eqn{j \in N(i)} and \eqn{N(i)} indexes the sites that are neighbors of the \eqn{i^{th}} site.
 #' 
 #' With the Gaussian probability distribution,
 #' \deqn{
@@ -75,9 +75,9 @@
 #' }
 #' where \eqn{\tau_i} is a scale parameter and \eqn{\mu_i} may contain covariates or simply the intercept.
 #'
-#' The covariance matrix of the CAR model contains two parameters: \code{car_rho} (\eqn{\rho}), which controls the degree of spatial autocorrelation, and the scale parameter, \code{car_scale} (\eqn{\tau}). The range of permissible values for \eqn{\rho} depends on the specification of \eqn{\boldsymbol C} and \eqn{\boldsymbol M}; for specification options, see \code{\link[geostan]{prep_car_data}} and Cressie and Wikle (2011, pp. 184-188) or Donegan (2021).
+#' The covariance matrix of the CAR model contains two parameters: \eqn{\rho} (\code{car_rho}) which controls the kind (positive or negative) and degree of spatial autocorrelation, and the scale parameter \eqn{\tau} (\code{car_scale}). The range of permissible values for \eqn{\rho} depends on the specification of \eqn{\boldsymbol C} and \eqn{\boldsymbol M}; for specification options, see \link[geostan]{prep_car_data} and Cressie and Wikle (2011, pp. 184-188) or Donegan (2021).
 #' 
-#' Further details of the models and results depend on the \code{family} argument, as well as on the particular CAR specification chosen (see \link[geostan]{prep_car_data}).
+#' Further details of the models and results depend on the \code{family} argument, as well as on the particular CAR specification chosen (from \link[geostan]{prep_car_data}).
 #'
 #' ###  Auto-Gaussian
 #'
@@ -85,15 +85,15 @@
 #' \deqn{
 #'  y \sim Gauss( \mu, (I - \rho C)^{-1} M),
 #' }
-#' where \eqn{\mu} is the mean vector (with intercept, covariates, etc.), \eqn{C} is a spatial connectivity matrix, and \eqn{M} is a known diagonal matrix containing the conditional variances \eqn{\tau_i^2}. `C` and `M` are provided by \code{\link[geostan]{prep_car_data}}.
[…]
