+++ R/stan_sar.R
@@ -73,7 +73,7 @@
 #' 
 #' Most often, this model is applied directly to observations (referred to below as the auto-Gaussian model). The SAR model can also be applied to a vector of parameters inside a hierarchical model. The latter enables spatial autocorrelation to be modeled when the observations are discrete counts (e.g., disease incidence data).
 #'
-#' A note on terminology: the spatial statistics literature conceptualizes the simultaneously-specified spatial autoregressive model (SAR) in relation to the conditionally-specified spatial autoregressive model (CAR) (see \code{\link[geostan]{stan_car}}) (see Cliff and Ord 1981). The spatial econometrics literature, by contrast, refers to the simultaneously-specified spatial autoregressive (SAR) model as the spatial error model (SEM), and they contrast the SEM with the spatial lag model (which contains a spatially-lagged dependent variable on the right-hand-side of the regression equation) (see LeSage 2014). 
+#' A note on terminology: the spatial statistics literature conceptualizes the simultaneously-specified spatial autoregressive model (SAR) in relation to the conditionally-specified spatial autoregressive model (CAR) (see \link[geostan]{stan_car}) (see Cliff and Ord 1981). The spatial econometrics literature, by contrast, refers to the simultaneously-specified spatial autoregressive (SAR) model as the spatial error model (SEM), and they contrast the SEM with the spatial lag model (which contains a spatially-lagged dependent variable on the right-hand-side of the regression equation) (see LeSage 2014). 
 #' 
 #' ###  Auto-Gaussian
 #'
@@ -89,9 +89,9 @@
 #' \phi = \rho W (y - \mu)
 #' }
 #' 
-#' This term can be extracted from a fitted auto-Gaussian model using the \code{\link[geostan]{spatial}} method.
+#' This term can be extracted from a fitted auto-Gaussian model using the \link[geostan]{spatial} method.
 #'
-#' When applied to a fitted auto-Gaussian model, the \code{\link[geostan]{residuals.geostan_fit}} method returns 'de-trended' residuals \eqn{R} by default. That is,
+#' When applied to a fitted auto-Gaussian model, the \link[geostan]{residuals.geostan_fit} method returns 'de-trended' residuals \eqn{R} by default. That is,
 #' \deqn{
 #' R = y - \mu - \rho W (y - \mu).
 #' }
@@ -106,7 +106,7 @@
 #' \lambda \sim Gauss(\mu, \Sigma)   \\
 #' \Sigma = \sigma^2 (I - \rho W)^{-1}(I - \rho W')^{-1}.
 #' }
-#' If the raw outcome consists of a rate \eqn{\frac{y}{p}} with observed counts \eqn{y} and denominator {p} (often this will be the size of the population at risk), then \eqn{O=log(p)} is the log of the denominator (the offset term).
+#' If the raw outcome consists of a rate \eqn{\frac{y}{p}} with observed counts \eqn{y} and denominator {p} (often this will be the size of the population at risk), then the offset term \eqn{O=log(p)} is the log of the denominator.
 #' 
 #' This is often written (equivalently) as:
 #' \deqn{
@@ -115,7 +115,7 @@
 #' \Sigma = \sigma^2 (I - \rho W)^{-1}(I - \rho W')^{-1}
 #' }
 #'
-#' For Poisson models, the \code{\link[geostan]{spatial}} method returns the parameter vector \eqn{\phi}.
+#' For Poisson models, the \link[geostan]{spatial} method returns the parameter vector \eqn{\phi}.
 #'
[…]
