### 2022-09-18 3a536df
cleanup text, add table
