+++ paper.md
@@ -5,7 +5,7 @@ tags:
 - spatial data
 - survey data
 - Bayesian inference
-date: "13 July, 2022"
+date: "17 September, 2022"
 output: pdf_document
 authors:
 - name: Connor Donegan
@@ -26,15 +26,15 @@ This paper introduces **geostan**, an `R` package for analyzing spatial data usi
 
 # Statement of need
 
-The distinguishing characteristic of spatial data is that maps of the data typically contain moderate to strong spatial patterns, or spatial autocorrelation, which reduces effective sample size (ESS), and renders many standard statistical tests inapplicable [@student_1914; @clifford_1989]. It is common for social variables to contain shared components of their spatial patterns, which inflates measures of association [@griffith_2019]. Spatial patterns are often of direct interest---disease mapping studies are concerned primarily with understanding how disease or mortality risk vary over social space.
+The distinguishing characteristic of spatial data is that maps of the data typically contain moderate to strong spatial patterns, or spatial autocorrelation, which reduces effective sample size (ESS) and renders many standard statistical methods inappropriate [@student_1914; @clifford_1989]. In addition, spatial patterns are often of direct interest---for example, disease mapping studies are concerned primarily with understanding how disease or mortality risk vary over space.
 
 A major challenge for spatial analysis is data quality, particularly for researchers using survey-based covariates. A single spatial analysis may use dozens, or even thousands, of error-laden survey estimates. Sampling error in ACS estimates is often substantial in magnitude and socially patterned [@folch_2016; @donegan_2021], which can have real consequences on communities and service providers [@bazuin_2013]. Spatial ME models are required to avoid ME biases and unwarranted levels of confidence in results.
 
-Existing R packages with spatial modeling functions include **spatialreg** [@bivand_2015], **INLA** [@rue_2009], **ngspatial** [@hughes_2020], **CARBayes** [@lee_2013], **nimble** [@valpine_2017]. Custom spatial models can be built using **rstan** [@stan_2022] or **nimble**, including spatial ME models, but this requires specialized programming and statistical skills. None of the other packages currently offer the spatial ME models that survey-based covariates require.
+Existing R packages with spatial modeling functions include **spatialreg** [@bivand_2015], **INLA** [@rue_2009], **ngspatial** [@hughes_2020], **CARBayes** [@lee_2013], **nimble** [@valpine_2017]. Custom spatial models can be built using **rstan** [@stan_2022] or **nimble**, including spatial ME models, but this requires specialized programming and statistical skills. **geostan** fills two gaps in this software landscape. First, **geostan** offers spatial ME models that are appropriate for survey-based covariates. Second, **geostan** provides spatial model diagnostic functions that make it easy for users to evaluate model results even if they are unfamiliar with Markov chain Monte Carlo (MCMC).
 
 # Functionality
 
-**geostan** provides convenient tools for a full spatial analysis workflow for spatial regression and disease mapping purposes. It provides tools for spatial data visualization, construction of spatial weights matrices (building upon `spdep` [@bivand_2013]), spatial ME models, models for censored count data, and multiple types of spatial statistical models for continuous and discrete data types. 
+**geostan** provides tools for spatial data visualization, construction of spatial weights matrices (building upon `spdep` [@bivand_2013]), spatial ME models, models for censored count data, and multiple types of spatial statistical models for continuous and discrete data types. 
 
[…]
