+++ paper.bib
@@ -45,6 +45,23 @@
   address={Los Angeles}
 }
 
+@article{student_1914,
+  title={The elimination of spurious correlation due to position in time and space},
+  author={{``Student'' [W.S. Gausset]}},
+  journal={Biometrika},
+  volume={10},
+  year={1914},
+  pages={179-180}
+}
+
+@Manual{hughes_2020,
+    title = {ngspatial: Fitting the Centered Autologistic and Sparse Spatial Generalized Linear Mixed Models for Areal Data},
+    author = {John Hughes and Xiaohui Cui},
+    year = {2020},
+    address = {Frederick, MD},
+    note = {R package version 1.2-2},
+  }
+
 @book{cliff_1981,
   title={Spatial Processes: Models and Applications},
   publisher={Pion},
@@ -205,6 +222,15 @@
   year={2016},
   publisher={Sage Publications Sage UK: London, England},
   doi={10.1177/0962280216660421}
+}
+
+@article{lesage_2014,
+  title={What regional scientists need to know about spatial econometrics},
+  author={LeSage, James P},
+  journal={The Review of Regional Studies},
+  year={2014},
+  volume={44},
+  pages={13--32}
 }
 
 @article{morris_2019,
