### 2022-09-19 1fa7365
add SAR, update fitted method
