+++ paper.bib
@@ -51,11 +51,12 @@
   journal={Biometrika},
   volume={10},
   year={1914},
+  doi={10.1093/biomet/10.1.179},
   pages={179-180}
 }
 
 @Manual{hughes_2020,
-    title = {ngspatial: Fitting the Centered Autologistic and Sparse Spatial Generalized Linear Mixed Models for Areal Data},
+    title = {{ngspatial}: Fitting the Centered Autologistic and Sparse Spatial Generalized Linear Mixed Models for Areal Data},
     author = {John Hughes and Xiaohui Cui},
     year = {2020},
     address = {Frederick, MD},
@@ -154,6 +155,10 @@
   author={Logan, John R and Bauer, Cici and Ke, Jun and Xu, Hongwei and Li, Fan},
   journal={Geographical Analysis},
   year={2019},
+  volume={52},
+  number={3},
+  pages={325--350},
+  doi={10.1111/gean.12215},
   publisher={Wiley Online Library}
 }
 
@@ -252,6 +257,7 @@
   number={2},
   pages={389--402},
   year={2019},
+  doi={10.1111/rssa.12378},
   publisher={Wiley Online Library}
 }
 
@@ -280,6 +286,7 @@
     journal = {Journal of the Royal Statistical Society B},
     year = {2009},
     volume = {71},
+    doi={10.1111/j.1467-9868.2008.00700.x},
     pages = {319--392},
   }
 
