### 2022-10-14 #4716 @becarioprecario
I have read the paper now. These are some comments that perhaps require changes in the main paper:

* Authorship. The geostan package has been vdeveloped by two authors but only one appears listed in the paper. Is there any particular reason for this?
* BayesX is also an R package to fit Bayesian spatial models that should be mentioned.
* Lines 55-59. Does the package relies on other R packages to compute these statistics and plots?
* Line 74. If there is a model that will be included in a forthcoming version I would wait to have these models implemented to actually publish the paper. It seems that the SAR model is already there, so the paper can be updated and this sentence rephrased.
* Line 81. INLA can also include user-defined models
* Citation format. When the references are included in a sentence which is between parentheses I believe that they are written using the wrong formant, e.g., "(... Author1 et al. (2022) ...)" instead of "(... Autho1 et al., 2022, ...)"
