+++ paper.bib
@@ -1,3 +1,30 @@
+ @Manual{bates_2022,
+    title = {Matrix: Sparse and Dense Matrix Classes and Methods},
+    author = {Douglas Bates and Martin Maechler and Mikael Jagan},
+    year = {2022},
+    note = {R package version 1.5-1/r3641},
+    url = {https://R-Forge.R-project.org/projects/matrix/},
+  }
+
+  @Article{umlauf_2015,
+    title = {Structured Additive Regression Models: An {R} Interface to {BayesX}},
+    author = {Nikolaus Umlauf and Daniel Adler and Thomas Kneib and Stefan Lang and Achim Zeileis},
+    journal = {Journal of Statistical Software},
+    year = {2015},
+    volume = {63},
+    number = {21},
+    pages = {1--46},
+    url = {https://www.jstatsoft.org/v63/i21/},
+  }
+
+  @Manual{belitz_2022,
+    title = {{BayesX}: Software for {B}ayesian Inference in Structured Additive Regression Models},
+    author = {Christiane Belitz and Andreas Brezger and Thomas Kneib and Stefan Lang and Nikolaus Umlauf},
+    year = {2022},
+    note = {Version 1.1},
+    url = {https://www.uni-goettingen.de/de/bayesx/550513.html},
+  }
+
 
 @article{donegan_2022,
   title={Building spatial conditional autoregressive (CAR) models in the {S}tan programming language},
@@ -310,6 +337,22 @@
     year = {2017},
     author = {Perry {de Valpine} and Daniel Turek and Christopher Paciorek and Cliff Anderson-Bergman and Duncan {Temple Lang} and Ras Bodik},
     doi = {10.1080/10618600.2016.1172487},
+  }
+
+@Manual{gabry_2020,
+    title = {rstantools: Tools for Developing R Packages Interfacing with 'Stan'},
+    author = {Jonah Gabry and Ben Goodrich and Martin Lysy},
+    year = {2020},
+    note = {https://mc-stan.org/rstantools/, https://discourse.mc-stan.org/},
+  }
+
+  @Manual{gabry_2018,
+    title = {shinystan: Interactive Visual and Numerical Diagnostics and Posterior
+Analysis for Bayesian Models},
+    author = {Jonah Gabry},
+    year = {2018},
+    note = {R package version 2.5.0},
+    url = {https://CRAN.R-project.org/package=shinystan},
   }
 
 
