+++ paper.md
@@ -22,7 +22,7 @@ affiliations:
 
 Analyses of data collected across areal units, such as census tracts and states, are now ubiquitous in the social and health sciences. Data sources include surveys (especially large government-back surveys like the US Census Bureau's American Community Survey (ACS)), vital statistics systems, and disease registries (particularly cancer registries). These data sources can provide crucial information about population health and socio-economic outcomes, but many standard (non-spatial) statistical methods and workflows are either not applicable to spatial data or require adjustment [@cressie_2015; @haining_2020].
 
-This paper introduces **geostan**, an `R` package for analyzing spatial data using Bayesian inference. The primary focus of the package is areal data for socio-economic and health research. The package provides tools for a complete workflow for spatial regression and disease mapping, and has unique spatial measurement error (ME) models suitable for researchers using American Community Survey (ACS) estimates as covariates [@donegan_2021].
+This paper introduces **geostan**, an `R` package for analyzing spatial data using Bayesian inference. The primary focus of the package is areal data for socio-economic and health research. The package provides tools for a complete workflow for spatial regression and disease mapping, and has unique spatial measurement error (ME) models suitable for researchers using ACS estimates as covariates [@donegan_2021].
 
 # Statement of need
 
@@ -30,13 +30,13 @@ The distinguishing characteristic of spatial data is that maps of the data typic
 
 A major challenge for spatial analysis is data quality, particularly for researchers using survey-based covariates. A single spatial analysis may use dozens, or even thousands, of error-laden survey estimates. Sampling error in ACS estimates is often substantial in magnitude and socially patterned [@folch_2016; @donegan_2021], which can have real consequences on communities and service providers [@bazuin_2013]. Spatial ME models are required to avoid ME biases and unwarranted levels of confidence in results.
 
-Existing R packages with spatial modeling functions include **spatialreg** [@bivand_2015], **INLA** [@rue_2009], **ngspatial** [@hughes_2020], **CARBayes** [@lee_2013], **nimble** [@valpine_2017]. Custom spatial models can be built using **rstan** [@stan_2022] or **nimble**, including spatial ME models, but this requires specialized programming and statistical skills. **geostan** fills two gaps in this software landscape. First, **geostan** offers spatial ME models that are appropriate for survey-based covariates. Second, **geostan** provides spatial model diagnostic functions that make it easy for users to evaluate model results even if they are unfamiliar with Markov chain Monte Carlo (MCMC).
+Existing R packages with spatial modeling functions include **spatialreg** [@bivand_2015], **INLA** [@rue_2009], **ngspatial** [@hughes_2020], **BayesX** [@belitz_2022; @umlauf_2015], **CARBayes** [@lee_2013], **nimble** [@valpine_2017]. Custom spatial models can be built using **rstan** [@stan_2022], **INLA*, and **nimble**, including spatial ME models, but this requires specialized programming and statistical skills. **geostan** fills two gaps in this software landscape. First, **geostan** offers spatial ME models that are appropriate for survey-based covariates. Second, **geostan** provides spatial model diagnostic functions that make it easy for users to evaluate model results even if they are unfamiliar with Markov chain Monte Carlo (MCMC).
 
 # Functionality
 
-**geostan** provides tools for spatial data visualization, construction of spatial weights matrices (building upon `spdep` [@bivand_2013]), spatial ME models, models for censored count data, and multiple types of spatial statistical models for continuous and discrete data types. 
[…]
