+++ paper.bib
@@ -25,6 +25,14 @@
     url = {https://www.uni-goettingen.de/de/bayesx/550513.html},
   }
 
+ @Book{wickham_2016,
+    author = {Hadley Wickham},
+    title = {ggplot2: Elegant Graphics for Data Analysis},
+    publisher = {Springer-Verlag New York},
+    year = {2016},
+    isbn = {978-3-319-24277-4},
+    url = {https://ggplot2.tidyverse.org},
+  }
 
 @article{donegan_2022,
   title={Building spatial conditional autoregressive (CAR) models in the {S}tan programming language},
