### 2022-10-21 #4716 @ConnorDonegan
Thank you again @becarioprecario for your review.

 - authorship: I'm the only package author, although there are two contributors to the code base. Some of the Stan code that I use is from the published work of others, including contributor Mitzi Morris (ICAR models) and the regularized horseshoe prior used in the `stan_esf` model (from a paper by Pirronen and Vehtari). I do cite their work in the documentation. (I think that answers the question, but I'm eager to give credit in all appropriate ways.)
 - I now cite the **R2BayesX** Rackage
 - The spatial auto-correlation statistics do not rely on any outside packages. However, the package uses `spdep` and the `Matrix` package to create the connectivity matrices. I have revised the **Functionality** section to clarify how `spdep` is used and add reference to `Matrix`. Graphics are created using `ggplot2`, now cited.
 - The text now reflects that the SAR models are currently available.
 - I now cite **INLA** as among the packages that provide extensive customization. I believe this includes the ability to specify spatial ME models very similar to those in **geostan** (a spatial model as the prior distribution on the latent covariate, with user-provided measurement error variances.) @becarioprecario you may be able to correct me if I'm wrong about that.
 - After revisions, there is one citation that sits within parentheses (line 80). I believe it is done correctly, and will defer to @jbytecode
