### 2022-10-23 #4716 @becarioprecario
Hi,

Thanks for your review.

  *   I now cite INLA as among the packages that provide extensive customization. I believe this includes the ability to specify spatial ME models very similar to those in geostan (a spatial model as the prior distribution on the latent covariate, with user-provided measurement error variances.) @becarioprecario<https://github.com/becarioprecario> you may be able to correct me if I'm wrong about that.

Yes. INLA allows the user to define latent effects by specifying the distribution of these effects,

  *   After revisions, there is one citation that sits within parentheses (line 80). I believe it is done correctly, and will defer to @jbytecode<https://github.com/jbytecode>

There is an extra bracket at the end that must be removed.

Best,

Virgilio
