### 2022-10-25 #4716 @becarioprecario
Hi,

I just pointed out a typo in my last e-mail. Other than that, I think that the paper is fine.

Best,

Virgilio
