### 2022-10-30 ConnorDonegan/geostan#3 @wcjochem
JOSS Review: Performance claims
The manuscript states that the CAR model is efficient for fitting models to county-level data for the entire U.S. on a laptop (lines 83-86). Do you have a worked example for users that demonstrates that claim? Alternatively, please provide a reference in the manuscript to a paper where the model is used at that scale to provide evidence.

https://github.com/openjournals/joss-reviews/issues/4716
