### 2022-10-30 ConnorDonegan/geostan#4 @wcjochem
JOSS Review: Contributing guidelines
Please add clear guidance for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

https://github.com/openjournals/joss-reviews/issues/4716
