### 2022-10-30 ConnorDonegan/geostan#5 @wcjochem
JOSS Review: Documentation of `geostan_fit` methods
Overall I found the functions to be adequately documented throughout; however, I think it could be improved if you could split the documentation for "`geostan_fit` methods" into separate pages for plotting/printing, extracting, predicting. Currently these are contained in a single documentation page and are indexed under one topic. These functions cover a wide range steps and it would be easier to find that documentation and I think it would make it more readable for users to have separate pages.

https://github.com/openjournals/joss-reviews/issues/4716
