### 2022-10-30 #4716 @wcjochem
@ConnorDonegan and @jbytecode, thanks for the opportunity to review this work. This package is an impressive accomplishment and clearly you've put a lot of thought and work into it. I have been able to build (and test) and run through the code and models and review the manuscript. Overall it seems well done. I have a raised a few questions in https://github.com/ConnorDonegan/geostan/issues (and linked above) for your consideration. When those are addressed, I can complete my checklist.
