### 2022-11-01 #4716 @ConnorDonegan
Many thanks for your review @wcjochem. I think the suggestion to break the `geostan_fit` methods documentation into a number of pages is a good idea. I'll do that and respond to your other comments, I think I can have that done by the middle of next week.

On the missing citation to support the performance claim (that the CAR models can work with all US county data interactively), I'll add a direct reference to my OSF preprint on `Building spatial CAR models'. The claim is a reference to the demonstration analysis in that paper, as well as the analysis of sampling efficiency in it. Thanks for catching that.
