### 2022-11-06 ConnorDonegan/geostan#3 @ConnorDonegan
Thanks @wcjochem, I'm adding a reference to the OSF pre-print "Building spatial conditional auto-regressive models in the Stan programming language" (https://osf.io/3ey65/) which has a demonstration analysis with US county data and also a study of sampling efficiency with head-to-head comparison between my Stan functions and NIMBLE's `dcar proper`.
