### 2022-11-06 ConnorDonegan/geostan#4 @ConnorDonegan
I'm now reading up on how to write a nice `CONTRIBUTOR.md` file (<https://opensource.guide/building-community/>), and I'm adding the following section to the paper:

## User support and discussion

In addition to consulting the package website and vignettes (<https://connordonegan.github.io/geostan>), users are encouraged to engage with **geostan** through the package github page (<https://github.com/connordonegan/geostan>). The "Discussions" tab is the best place for users to seek support and ask questions. Any issues with the software or suggestions for new features may also be proposed there or on the "Issues" tab. Contributions to the software can also be proposed by making a pull request to the github page.
