### 2022-11-07 ConnorDonegan/geostan#5 @ConnorDonegan
Thanks, this is a really helpful suggestion. I see its looking cluttered right now.

I'm breaking the `geostan_fit` methods documentation into the following pages:

 - Extract residuals, fitted values, or spatial trend
 - predict
 - print or plot model results
 - extract MCMC samples (as.matrix, as.data.frame, as.array)

I'd like to push this and the other changes to CRAN just after the publication is finished and can be cited.
