### 2022-11-07 edb90a1
clarify methods doc
