+++ R/geostan_fit-methods.R
@@ -1,62 +1,24 @@
-#' geostan_fit methods
+#' print or plot a fitted geostan model
 #'
-#' @description Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.
+#' @description Print a summary of model results to the R console, or plot posterior distributions of model parameters.
 #'
-#' @param object A fitted model object of class \code{geostan_fit}.
 #' @param x A fitted model object of class \code{geostan_fit}.
-#' @param summary Logical; should the values be summarized with the mean, standard deviation and quantiles (\code{probs = c(.025, .2, .5, .8, .975)}) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.
-#' @param newdata A data frame in which to look for variables with which to predict, presumably for the purpose of viewing marginal effects. Note that if the model formula includes an offset term, `newdata` must contain a column with the appropriate name for the offset, even though the values will be ignored (you may set all values to 1); you must use the `alpha` argument to include any additional terms. Note also that any spatially-lagged covariate terms will be ignored if they were provided using the `slx` argument. If covariates in the model were centered using the `centerx` argument, the `predict.geostan_fit` method will automatically center the predictors in `newdata` internally using the values stored in `fit$x_center`. If `newdata` is missing, user arguments will be passed to the `fitted.geostan_fit` method to return the fitted values of the model.
-#' 
-#' @param alpha A single numeric value or a numeric vector with length equal to `nrow(newdata)`; `alpha` serves as the intercept in the linear predictor. The default is to use the posterior mean of the intercept. Even if \code{type = "response"}, this needs to be provided on the scale of the linear predictor. See `Details` for additional information.
-#'
-#' @param center May be a vector of numeric values or a logical scalar to pass to \code{\link[base]{scale}}. Defaults to using `object$x_center`. If the model was fit using `centerx = TRUE`, then covariates were centered and their mean values are stored in `object$x_center` and the `predict` method will use them to automatically center `newdata`; if the model was fit with `centerx = FALSE`, then `object$x_center = FALSE` and `newdata` will not be centered. 
-#' 
-#' @param type By default, results from `predict` are on the scale of the linear predictor (`type = "link")`). The alternative (`type = "response"`) is on the scale of the response variable. For example, the default return values for a Poisson model are log-rates, and using `type = "response"` will return the rates (by exponentiating the log-rates).
-#' 
+#' @param digits number of digits to print
+#' @param probs Argument passed to \code{quantile}; which quantiles to calculate and print.
 #' 
 #' @param pars parameters to include; a character string (or vector) of parameter names.
[…]
