+++ man/geostan_fit.Rd
@@ -1,163 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/geostan_fit-methods.R
-\name{print.geostan_fit}
-\alias{print.geostan_fit}
-\alias{plot.geostan_fit}
-\alias{as.matrix.geostan_fit}
-\alias{as.data.frame.geostan_fit}
-\alias{as.array.geostan_fit}
-\alias{residuals.geostan_fit}
-\alias{fitted.geostan_fit}
-\alias{spatial}
-\alias{spatial.geostan_fit}
-\alias{predict.geostan_fit}
-\title{geostan_fit methods}
-\usage{
-\method{print}{geostan_fit}(
-  x,
-  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
-  digits = 3,
-  pars = NULL,
-  ...
-)
-
-\method{plot}{geostan_fit}(x, pars, plotfun = "hist", fill = "steelblue4", ...)
-
-\method{as.matrix}{geostan_fit}(x, ...)
-
-\method{as.data.frame}{geostan_fit}(x, ...)
-
-\method{as.array}{geostan_fit}(x, ...)
-
-\method{residuals}{geostan_fit}(object, summary = TRUE, rates = TRUE, detrend = TRUE, ...)
-
-\method{fitted}{geostan_fit}(object, summary = TRUE, rates = TRUE, trend = TRUE, ...)
-
-spatial(object, summary = TRUE, ...)
-
-\method{spatial}{geostan_fit}(object, summary = TRUE, ...)
-
-\method{predict}{geostan_fit}(
-  object,
-  newdata,
-  alpha = mean(as.matrix(object, pars = "intercept")),
-  center = object$x_center,
-  summary = TRUE,
-  type = c("link", "response"),
-  ...
-)
-}
-\arguments{
-\item{x}{A fitted model object of class \code{geostan_fit}.}
-
-\item{probs}{Argument passed to \code{quantile}; which quantiles to calculate and print.}
-
-\item{digits}{number of digits to print}
-
-\item{pars}{parameters to include; a character string (or vector) of parameter names.}
-
-\item{...}{additional arguments.}
-
-\item{plotfun}{Argument passed to \code{rstan::plot}. Options include histograms ("hist"), MCMC traceplots ("trace"), and density plots ("dens"). Diagnostic plots are also available such as Rhat statistics ("rhat"), effective sample size ("ess"), and MCMC autocorrelation ("ac").}
-
-\item{fill}{fill color for histograms and density plots.}
-
-\item{object}{A fitted model object of class \code{geostan_fit}.}
-
-\item{summary}{Logical; should the values be summarized with the mean, standard deviation and quantiles (\code{probs = c(.025, .2, .5, .8, .975)}) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.}
-
-\item{rates}{For Poisson and Binomial models, should the fitted values be returned as rates, as opposed to raw counts? Defaults to \code{TRUE}.}
-
-\item{detrend}{For auto-normal models (CAR and SAR models with Gaussian likelihood only); if \code{detrend = TRUE}, the implicit spatial trend will be removed from the residuals. The implicit spatial trend is \code{Trend = rho * C \%*\% (Y - Mu)} (see \link[geostan]{stan_car} or \link[geostan]{stan_sar}). I.e., \code{resid = Y - (Mu + Trend)}.}
-
[…]
