### 2022-11-07 54b172f
streamline text
