### 2022-11-07 #4716 @ConnorDonegan
Thanks again @wcjochem for your review. I think you've already seen a notification for my responses to your comments over on the issues tab for the project github page.

I'll also leave a note of the changes here for everyone.

### Performance claims
> The manuscript states that the CAR model is efficient for fitting models to county-level data for the entire U.S. on a laptop (lines 83-86). Do you have a worked example for users that demonstrates that claim? Alternatively, please provide a reference in the manuscript to a paper where the model is used at that scale to provide evidence.

I've added a reference to the OSF pre-print "Building spatial conditional auto-regressive models in the Stan programming language" (https://osf.io/3ey65/) which has a demonstration analysis with US county data and also a study of sampling efficiency with head-to-head comparison between my Stan functions and NIMBLE's dcar proper.

### Contributor guidelines
> Please add clear guidance for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

I'm glad to be adding the following section to the paper:

**User support and discussion**

In addition to consulting the package website and vignettes (<https://connordonegan.github.io/geostan>), users are encouraged to engage with **geostan** through the package github page (<https://github.com/connordonegan/geostan>). The github "Issues" and "Discussions" tabs are the best place for users to seek support, ask questions, report bugs, or suggest new features. Contributions to the software can also be proposed by making a pull request to the github page.

### `geostan_fit` methods documentation
> Overall I found the functions to be adequately documented throughout; however, I think it could be improved if you could split the documentation for "geostan_fit methods" into separate pages for plotting/printing, extracting, predicting. Currently these are contained in a single documentation page and are indexed under one topic. These functions cover a wide range steps and it would be easier to find that documentation and I think it would make it more readable for users to have separate pages.

This is a really helpful suggestion. I see it was looking cluttered.

I've broken the `geostan_fit` methods documentation into the following pages:

- [Extract residuals, fitted values, or spatial trend](https://connordonegan.github.io/geostan/reference/resid_geostan_fit.html)
- [predict](https://connordonegan.github.io/geostan/reference/predict.geostan_fit.html)
- [print or plot model results](https://connordonegan.github.io/geostan/reference/print_geostan_fit.html)
- [extract MCMC samples (as.matrix, as.data.frame, as.array)](https://connordonegan.github.io/geostan/reference/samples_geostan_fit.html)

This change to the documentation is now reflected on the website.
