### 2022-11-08 #4716 @jbytecode
@ConnorDonegan - the item

`Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support`

belongs to the documentation rather than the software paper, I think the most suitable place to host them is the repository - not the paper. @wcjochem makes the final decision, of course.
