### 2022-11-08 #4716 @ConnorDonegan
I would not object to that, @jbytecode. I've also added this to the README for the github repo and to the main website.

The website/readme now has this:

### Support
All functions and methods are documented (with examples) on the website [reference](https://connordonegan.github.io/geostan/reference/index.html) page. See the package [vignettes](https://connordonegan.github.io/geostan/articles/index.html) for more on exploratory spatial data analysis and spatial modeling.

To ask questions, report a bug, or discuss ideas for improvements or new features please visit the [Issues](https://github.com/ConnorDonegan/geostan/issues) page, start a [discussion](https://github.com/ConnorDonegan/geostan/discussions), or submit a [pull request](https://github.com/ConnorDonegan/geostan/pulls).
