### 2022-11-09 #4716 @wcjochem
Thanks @ConnorDonegan for these revisions. I agree with @jbytecode the contributing guidelines are most appropriate in the README/website. I think you have addressed all my issues, so I'm happy to complete my review now. Thanks!
