+++ paper.md
@@ -83,10 +83,6 @@ A set of functions for working with model results conveniently extract fitted va
 
 ME models can be added to any **geostan** model. These are models for covariates measured with error, particularly small-area survey estimates with standard errors. The ME models treat the true covariate values as unknown parameters or latent variables, which are assigned a spatial CAR prior model. Users provide the scale of observational uncertainty or ME (e.g., survey standard errors) as data [@donegan_2021; cf. @bernardinelli_1997; @xia_1998; @kang_2009; @logan_2019]. All uncertain inferences from the ME models are automatically propagated throughout the regression or disease mapping model, and graphical diagnostics are provided for evaluating results of spatial ME models.
 
-## User support and discussion
-
-In addition to consulting the package website and vignettes (<https://connordonegan.github.io/geostan>), users are encouraged to engage with **geostan** through the package github page (<https://github.com/connordonegan/geostan>). The github "Issues" and "Discussions" tabs are the best place for users to seek support, ask questions, report bugs, or suggest new features. Contributions to the software can also be proposed by making a pull request to the github page.
-
 # Acknowledgements
 
 I am grateful for support this project received from Esri Inc. and from the Geography and Geospatial Information Sciences program at The University of Texas at Dallas.
