### 2022-11-09 #4716 @ConnorDonegan
@jbytecode I've now updated the paper.md and pdf file to no longer include the section on user support.
