+++ paper.bib
@@ -420,4 +420,13 @@ Analysis for Bayesian Models},
   doi={10.1111/j.1538-4632.1995.tb00338.x}
 }
 
+@Manual{rstats,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2021},
+    url = {https://www.R-project.org/},
+  }
+
 
