### 2022-11-10 #4716 @jbytecode
@ConnorDonegan - I have just sent a pull request that includes some corrections/additions to the manuscript and bibtex. Please review the changes and apply them if you are agreed with them. Thank you in advance.
