+++ paper.bib
@@ -145,7 +145,7 @@
 }
 
 @book{haining_2020,
-  title={Modelling Spatial and Spatio-Temporal Data: A Bayesian Approach},
+  title={Modelling Spatial and Spatio-Temporal Data: A {B}ayesian Approach},
   author={Haining, Robert P. and Li, Guangquan},
   publisher = {CRC Press},
   address={Boca Raton, FL, USA},
@@ -175,7 +175,7 @@
 }
 
 @article{folch_2016,
-   title={Spatial Variation in the Quality of American Community Survey Estimates},
+   title={Spatial Variation in the Quality of {A}merican {C}ommunity {S}urvey Estimates},
    author={Folch, David C. and Arribas-Bel, Daniel and Koschinsky, Julia and Spielman, Seth E.},
    year={2016},
    journal={Demography},
@@ -220,7 +220,7 @@
 }
 
 @article{xia_1998,
-  title={Spatio-temporal models with errors in covariates: mapping Ohio lung cancer mortality},
+  title={Spatio-temporal models with errors in covariates: mapping {O}hio lung cancer mortality},
   author={Xia, Hong and Carlin, Bradley P},
   journal={Statistics in Medicine},
   volume={17},
@@ -253,7 +253,7 @@
 }
 
 @article{riebler_2016,
-  title={An intuitive Bayesian spatial model for disease mapping that accounts for scaling},
+  title={An intuitive {B}ayesian spatial model for disease mapping that accounts for scaling},
   author={Riebler, Andrea and S{\o}rbye, Sigrunn H and Simpson, Daniel and Rue, H{\aa}vard},
   journal={Statistical methods in medical research},
   volume={25},
@@ -274,7 +274,7 @@
 }
 
 @article{morris_2019,
-  title={Bayesian hierarchical spatial models: Implementing the Besag York Molli{\'e} model in stan},
+  title={Bayesian hierarchical spatial models: Implementing the {B}esag {Y}ork {M}olli{\'e} model in {s}tan},
   author={Morris, Mitzi and Wheeler-Martin, Katherine and Simpson, Dan and Mooney, Stephen J and Gelman, Andrew and DiMaggio, Charles},
   journal={Spatial and Spatio-temporal Epidemiology},
   volume={31},
@@ -348,7 +348,7 @@
   }
 
 @Manual{gabry_2020,
-    title = {rstantools: Tools for Developing R Packages Interfacing with 'Stan'},
+    title = {rstantools: Tools for Developing {R} Packages Interfacing with '{S}tan'},
     author = {Jonah Gabry and Ben Goodrich and Martin Lysy},
     year = {2020},
     note = {https://mc-stan.org/rstantools/, https://discourse.mc-stan.org/},
@@ -356,7 +356,7 @@
 
   @Manual{gabry_2018,
     title = {shinystan: Interactive Visual and Numerical Diagnostics and Posterior
-Analysis for Bayesian Models},
+Analysis for {B}ayesian Models},
     author = {Jonah Gabry},
     year = {2018},
     note = {R package version 2.5.0},
