### 2022-11-10 #4716 @ConnorDonegan
Yeah---I'm going to drop that DOI. The journal article has been listed on site called SSRN, and it has a DOI. However, the journal article itself has no DOI associated with the 'real' publication. So this one will be missing a doi
