### 2022-11-10 #4716 @jbytecode
@ConnorDonegan - I have performed a DOI search using [this site](https://www.doi.org/) and it redirects to the [page](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2420725). So, it is okay, imo.
