### 2022-11-10 802ac4a
add SSRN doi to lesage
