+++ paper.bib
@@ -270,7 +270,8 @@
   journal={The Review of Regional Studies},
   year={2014},
   volume={44},
-  pages={13--32}
+  pages={13--32},
+  doi={10.2139/ssrn.2420725}
 }
 
 @article{morris_2019,
