### 2022-11-11 #4716 @tarleb
This looks like a pandoc bug. As a workaround, the `\label` should be moved directly below the `\caption` of the table.
